+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to Family Diagram are documented in this file.
 - Fixed broked "Add birth" and "Edit birth" buttons in person properties
 - Fixed person picker in add form not showing renamed people in auto-complete.
 - Can now use down/up arrows + enter/return to pick a person via autocomplete in add event form.
+- Visual timeline shows person names now.
 
 ## 2.1.16